Add ExitZone prefab and script Modify RecordFlagOrder.
ExitZone objects will trigger the endstates of the
level. The first ExitZone will start the second phase
of the current level. The second ExitZone will end the
level.

FlagManager.RecordFlagOrder has been modified to
check the length of the collectedFlagSequence list
against the number of flags that were instantiated.
If the length of collectedFlagSequence is equal to
(or is somehow greater than) the number of flags,
then it will deactivate the colliders in ExitZone00.
Curretnly, it prints a debug message.