Modify PlayerInteract, FlagProperties, FlagManager
Removed RaiseFlagSprite() call from PlayerInteract.
RaiseFlagSprite now moved to be called in
FlagProperties.SetIsPickedUp.

SetIsPickedUp now checks levelState and isPickedUp
before setting isPickedUp and calling RaiseFlagSprite.
If condition are met, then RecordFlagOrder or
CheckReverseOrder is called.

FlagManager has anew method, CheckReverseOrder. This
checks that the current flag that's being picked up
is the expected flag. If it is, then totalFlags is
decremented and then it returns true. If not, then
it returns false. If totalFlags reaches 0, then
GameManager.UnlockExitZone is called.