Add TriggerEndLevel Modified PlayerInteract
TriggerEndLevel added to GameManager. It will
trigger the end level screen. At the moment, it
just prints a debug message.

PlayerInteract.OnTriggerEnter2D has been modified.
The switch case for Exit00 now disables the
Exit00 game object, which causes a visual change
to the level. This way, the player knows something
has happened on reaching ExitZone00. There's also
a trophy graphic to encourage players to head to
that zone first.