Add Sprites for flag points
FlagProperties now holds a list containing
the list of sprites that represent the
numerical point value of the flag.

Added new method in FlagProperties,
UpdateFlagPointTextSprites, that updates
shows the corresponding sprite for the number.
This assumes that the number is only 2 digits
long and is an integer.

Imported sprites that represent the number.
Sprites are based on Thaleah_PixelFont.